% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runboth.R
\name{run.all}
\alias{run.all}
\title{Extreme value regression tests for range shifts}
\usage{
run.all(dataset, resp = "elev", n.pts = 10, OLE = FALSE,
  byRegion = NULL, latConv = FALSE)
}
\arguments{
\item{dataset}{A dataset frame containing columns "Year" and "Species" at a minimum as well as the response variable.}

\item{resp}{A variable within the dataset that is being tracked, such as elevation, or latitude. (If latitude is used, just note to convert it back to km manually, it's not currently automatic in here.)}

\item{n.pts}{Number of top-per-year points to use. Haven't added this functionality yet}

\item{OLE}{Distance-to-edge model, haven't added yet}

\item{byRegion}{The column to use as a random spatial effect to handle spatial heterogeneity (multiple gradients)}
}
\description{
Approach used in Carlson 2019 bioRxiv forthcoming. Can be done with maximum point per year, maximum n points, or the OLE ('distance-to-edge' model).
}
